Add Status Code option if needed when building an IA_PD option.
Make PrefixDelegation constructor with statusCode param visible for
testing, which is useful to inject the Status code in the response from
Server, e.g. inject an NoPrefixAvail status code in resposne to check
the client's behavior. And take the Status Code optino in IA_PD unless
it's not Success, otherwise, the absent of Status Code option in IA_PD
indicates a success resposne.

Bug: 317161875
Test: atest NetworkStackIntegrationTests
Change-Id: Ib67135c662e5cdf7ea644742503d4977f5e97fa8