Add an experiment flag for populating LinkAddress lifetime.
The flag is disabled by default, the same behavior as before, i.e. don't
populate the IP link address lifetime when user space sees the netlink
messages, and IpClient still calls Netd API to install an IPv4 address.
The refactor is guarded by a flag, which would be safer to rollout.

Bug: 129660498
Test: atest NetworkStackIntegrationTests
Test: atest NetworkStackRootTests
Change-Id: I5c7155e04423d84241144e3249969e86ee78987d