Fix flaky testPopulateLinkAddressLifetime.
The flakiness happened on the old devices such as sargo or redfin. Try
to reproduce the flaky test, looks like the dual-stack provisioning
takes ~3s to complete sometimes, DHCP handshake completes quickly but
IPv6 provisioning seems to take a bit longer to see the RTM_NEWADDR in
the user space. Increase the lifetime tolerance to 4s to defalky the
tests.

Bug: 322941589
Test: atest testPopulateLinkAddressLifetime --rerun-until-failure
Change-Id: I022bc07eb444f6c6c6b58f953bd782ea23e7267a