Add adb cmd support to pause and resume APF filter updates
This change is required to pause ApfFilter from installing newly updated
filter programs during integration tests.

Note: In its current state, there is a possible race between the
installation of a new filter program and a follow-up command to "pause".
While pause is executed on the IpClient handler, filter updates
as the result of a received RA circumvent the IpClient's handler. There
currently exists no synchronization mechanism between the two. That
being said, the "pause" command will be invoked via shell, so it seems
highly unlikely that this race could ever become a problem.
Nevertheless -- as a future TODO -- RA processing should be moved onto
the HandlerThread (which would also allow us to remove the locking
semantics in ApfFilter).

Test: adb shell cmd network_stack apf wlan0 status \
&& adb shell cmd network_stack apf wlan0 pause \
&& adb shell cmd network_stack apf wlan0 status \
&& adb shell cmd network_stack apf wlan0 resume \
&& adb shell cmd network_stack apf wlan0 status

Change-Id: If617530dcd09b36a2bd0879e00a0601fcd46b5be