Clean up population of LinkAddress deprecation and expiration times
I originally set out to fix a possible NullPointerException that was
introduced in aosp/2901458. I only noticed too late that this had
already been fixed in a later CL, so I ended up refactoring this code
for readability.

This change removes getDeprecationOrExpirationTime() which was hard to
read in favor of inlining the contained logic.

Test: TH
Change-Id: Icbef0d85336e36df340c3d591748d3846e36e19e