Remove the unnecessary lock from TcpSocketTracker
As the review feedback from aosp/2264841, TcpSocketTracker
should ever run on the NetworkMonitor handler thread, so add
a lock is unnecessary. The receiver registration can take
the handler from NetworkMonitor to receive the notification
via the handler thread. Update the design to register the
receiver in the default state of NetworkMonitor. This does
not impact the functionality as TcpSocketTracker does not
require to know doze mode status until first validation.

Bug: 268444709
Test: atest NetworkStackTests
Change-Id: I12e564b37eb432c74481cb6b3668184db1147020