Iterate through IpClients in reverse when using adb cmd apf
Tests often quickly create and destroy IpClients on the same interface.
This means that it is possible for an old IpClient instance to not be
fully garbage collected even after provisioning on a new IpClient has
succeeded. In this case, APF adb cmds could get a hold of the wrong
IpClient instance, since mIpClients is a List and new instances are
always appended to it.

This change reverses the list iteration, so that the most recent
IpClient instance is used in APF adb cmds.

Test: adb shell cmd network_stack apf wlan0 status
Bug: 331504289
Change-Id: I4a156327888851c2a9f44c9a8a55fd46a44ce1ef