Implement APF cmd support APIs in LegacyApfFilter
Unfortunately, it is not possible to consistently run tests
against the new ApfFilter. If an IpClient already exists (due to a
previous test not releasing its network request), enabling the
experiment flag before running ApfIntegrationTests is a no-op.

This change implements the required functionality to support APF adb
commands in LegacyApfFilter. (A nice side effect of which is that this
will be useful for debugging as well).

Test: TH
Bug: 333208995
Change-Id: I0a455e7290f55ec35e5a8be3a64c1822392401fd