ApfV6Generator - move overrides to Base class
These are APFv6 specific, so need to be in V6,
but (with current plans) there is no desire to change
these in APFv7/8+, so they should not be in the leaf
class but rather just in the common (to v6+) base class.

This is a simple cut'n'paste, just adjusting the return
type from 'public ApfV6Generator' to 'public final Type'.

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Icbdee2c5fa5b91974e6b758a4011be91fc8b66fd