Add PopulateLinkAddressLifetime annotation in IpClient integration test.
We've enabled the experiment flag that populates the LinkAddress
lifetime in setup by default, as a result, it triggers the
onLinkPropertiesChange callback whenever the lifetime changes. In order
no to mess up other test cases which doesn't require this flag, add a
new annotation class and assoicate it to test cases which need to test
the log of populating the link address lifetime.

Bug: 334567961
Test: atest NetworkStackRootTests
Change-Id: Ie36068fd14f92699d105b7341767a27e97d96a7d