ApfV4GeneratorBase: allow negative offsets in addLoad{8,16,32}Indexed
as these are relative offsets, so *not* (directly) packet offsets.

Found by Patrick and I when we tried to read u32 @-4 from
R1 = PACKET_SIZE.

For addLoad{8,16,32} the 'ofs' is an offset in the packet,
ie. 0 means first byte of the packet.

But for addLoad{8,16,32}Indexed the 'ofs' is an offset relative
to R1, so 'R1 + ofs' is the packet offset,
and 'ofs' itself is just a generic unsigned value.

addPacketOffset() is equivalent to addUnsigned(),
but we switch to addTwosCompUnsigned() instead,
so we can pass in negative offsets as well (in case
R1 is for example PACKET_SIZE)

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I976f7731b62b72f6577fcf5bef4237a548066927