Ignore the NUD failures from neighbor which is never reachable.
This CL attempts to ignore all critical NUD failure events from the
neighbor which is never ever reachable, i.e. user space never sees a
NUD_REACHABLE event for this neighbor, this may happen on address
resolution process (e.g. the state changes from null to NUD_FAILED) or
probing state (e.g. from NUD_PROBE to NUD_FAILED). Previously we've
implemented the logic of ignoring the incomplete neighbors, but it's
better to just ignore all NUD failures from the never reachable
neighbors.

So far we have a bug in IpReachabilityMonitor detecting the neighbor
lost event and triggering the onReachabilityFailure callback. For
example, we have two reachable default routers and one DNS server which
is never ever reachable, we got a NUD_FAILED event from the DNS server
first, then ignore the failure from it and also remove it from the list.
Later, we got another NUD_FAILED from one of default routers, given the
DNS server has been removed from watchlist and the code will remove the
default route after that, onReachabilityFailure will be triggered even
if the LP has another valid default route.

Instead of removing the neighbor from the watchlist, it's better to
ignore the neighbor when iterating the watchlist and skip the neighbor
if it's still never ever reachable on receiving the NUD_FAILUED event.

The new behavior is guarded by an experiment flag as usual, and disabled
by default, therefore, this should be a no-op change.

Bug: 329009735
Test: atest NetworkStackTests
Test: atest NetworkStackRootTests
Change-Id: I472ad0b6dc5ac350a35646b76c00da9f300a04a7